+++ src/main/java/com/dto/CreateClassDTO.java
@@ -1,11 +1,15 @@
 package com.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class CreateClassDTO {
+    @NotBlank(message = "Class name is required")
     private String className;
+
+    @NotBlank(message = "Section is required")
     private String section;
 }